Move enums to models package
